projects/meme/requirements.txt
asyncio
python-telegram-bot